I have changed a lot. I have changed things to be more BaseClassifierish and to work on OpenCL paltforms and not just GCD platforms (OS X).
In a bit I will commit again, with a lot of stuff removed. This will hopefully allow me to see how to make it to be something I can build (which will allow a lot of debugging).

I really just want to test the BaseClassifier right now, I will test the Interface when I use Eval and Classify.